Corrected some comments. Moved all the new functions except `create_display_plane_surface` to `PhysicalDevice`.
