+++ src/backend/empty/src/lib.rs
@@ -159,6 +159,25 @@ impl adapter::PhysicalDevice<Backend> for PhysicalDevice {
             ..Default::default()
         }
     }
+
+    fn enumerate_available_displays<'a>(&'a self)->Result<Vec<display::Display<'a,Backend>>,device::OutOfMemory> {unimplemented!();}
+
+    fn enumerate_compatible_planes<'a>(&self,_display: &display::Display<'a,Backend>)->Result<Vec<display::Plane<'a,Backend>>,device::OutOfMemory> {unimplemented!();}
+
+    fn enumerate_builtin_display_modes<'a>(&self,_display: &'a display::Display<'a,Backend>,)->Result<Vec<display::DisplayMode<'a,Backend>>,device::OutOfMemory> {unimplemented!();}
+
+    fn create_display_mode<'a>(
+        &self,
+        _display: &'a display::Display<'a,Backend>,
+        _resolution: (u32,u32),
+        _refresh_rate: u32
+    )->Result<display::DisplayMode<'a,Backend>,display::DisplayModeError> {unimplemented!();}
+
+    fn create_display_plane<'a>(
+        &self,
+        _display: &'a display::DisplayMode<'a,Backend>,
+        _plane: &'a display::Plane<'a,Backend>,
+    )->Result<display::DisplayPlane<'a,Backend>,device::OutOfMemory> {unimplemented!();}
 }
 
 /// Dummy command queue doing nothing.
@@ -1084,25 +1103,6 @@ impl hal::Instance<Backend> for Instance {
     }
 
     unsafe fn destroy_surface(&self, _surface: Surface) {}
-
-    fn enumerate_available_displays<'a>(&self,_adapter: &'a adapter::Adapter<Backend>)->Result<Vec<display::Display<'a,Backend>>,device::OutOfMemory> {unimplemented!();}
-
-    fn enumerate_compatible_planes<'a>(&self,_display: &display::Display<'a,Backend>)->Result<Vec<display::Plane<'a,Backend>>,device::OutOfMemory> {unimplemented!();}
-
-    fn enumerate_builtin_display_modes<'a>(&self,_display: &'a display::Display<'a,Backend>,)->Result<Vec<display::DisplayMode<'a,Backend>>,device::OutOfMemory> {unimplemented!();}
-
-    fn create_display_mode<'a>(
-        &self,
-        _display: &'a display::Display<'a,Backend>,
-        _resolution: (u32,u32),
-        _refresh_rate: u32
-    )->Result<display::DisplayMode<'a,Backend>,display::DisplayModeError> {unimplemented!();}
-
-    fn create_display_plane<'a>(
-        &self,
-        _display: &'a display::DisplayMode<'a,Backend>,
-        _plane: &'a display::Plane<'a,Backend>,
-    )->Result<display::DisplayPlane<'a,Backend>,device::OutOfMemory> {unimplemented!();}
 
      fn create_display_plane_surface(
         &self,
