+++ src/hal/src/common/mod.rs
@@ -1,0 +1,14 @@
+
+/*
+pub struct Extent
+{
+    width: u32,
+    height: u32
+}
+
+pub struct Origin
+{
+    x: i32,
+    y: i32
+}
+*/
